Build 4 groundwork: close out the Build 3 handoff
Everything HANDOFF.md left open, plus three bugs found while doing it.

Bugs
- `child:` / `name^:` / `name$:` answered differently depending on the
  rest of the query: the hydrated pass compared raw strings while the
  name-index pass ran the full normalization ladder, so adding `size:`
  to a query silently changed its hit set.
- The SRC-M23 match-mode flags never reached the daemon. `query.ts` sent
  all seven, `QueryRunParams` read all seven, and the Tauri command in
  between declared none of them — Tauri drops undeclared invoke
  arguments silently, so the whole `Search →` menu moved a checkmark and
  changed nothing.
- No setting has been persisting. `custom_commands` had a default and a
  settings panel but was missing from `ALLOWED_PATCH_KEYS`; `settings_set`
  rejects a patch on its first unknown key and `flush()` sends the entire
  state, so every save failed into a console warning.

The natural-sort opt-out now reaches the daemon, so `freally search`
honours it rather than only the client-side comparator.

Performance
- Ignore Punctuation / Whitespace seed from a parallel set of trigram
  postings over stripped names instead of walking the whole index. Built
  lazily on first use, so nobody pays for a mode they never turn on.
- One normalization ladder (`normalized`) replaces three drifting copies,
  and it borrows rather than allocating when a rung has nothing to do.
- Needles go through it once per query, not once per candidate row.
- Waveform bucketing carries its boundary instead of dividing per frame;
  the permission ledger dedupes through a set; `recent_searches` stops
  reassigning the settings root on every keystroke.

Duplication
- `allowed_patch_keys()` derives from the schema — this is what surfaced
  the settings bug.
- `portable::open_log`, `service::scan`, `app_data_root`,
  `resultsStore.hitById` replace their copies.
- A shared `<Modal>` for seven dialogs. It also fixes Escape, which never
  worked: the handler sat on a backdrop that could not receive keydown.
  Clears all six `a11y_interactive_supports_focus` warnings.

Freally Central is gone — vendored tree, submodule stanza, React island,
its dependencies, its locale strings, and its docs entries.

CI
- `src-tauri` is linted now; it was excluded from the workspace, so three
  clippy errors had been failing silently.
- nextest (100s to 37s warm), cancel-in-progress, `CARGO_INCREMENTAL=0`,
  cache written only from `main`, `--locked` throughout, frozen pnpm
  install, and the OS-agnostic gates on one runner.
- `ci-local.mjs` gained the src-tauri gate it was missing and runs its
  Rust and UI lanes concurrently: 133s of work in 89s.

+++ apps/freally-ui/src-tauri/src/commands/icons.rs
@@ -93,8 +93,8 @@ mod imp {
             }
             return None;
         }
-        let width = bmp.bmWidth as i32;
-        let height = bmp.bmHeight as i32;
+        let width = bmp.bmWidth;
+        let height = bmp.bmHeight;
         if width <= 0 || height <= 0 {
             unsafe {
                 let _ = DeleteObject(color.into());
@@ -156,7 +156,6 @@ mod imp {
     fn encode_wide(s: &str) -> Vec<u16> {
         s.encode_utf16().chain(std::iter::once(0)).collect()
     }
-
 }
 
 #[cfg(not(windows))]
